refactoring/move dimensions list related components into a subfolder (#211)
Changes include:

* rename Dimensions folder to DimensionsPanel
* move all components specific to the Dimensions list into the DimensionsPanel/List folder
* no functional changes
