+++ packages/app/src/components/DimensionsPanel/Dialogs/DataDimension/Detail.js
