+++ packages/app/src/components/DimensionsPanel/Dialogs/DataDimension/styles/Details.style.js
