+++ packages/app/src/components/DimensionsPanel/Dialogs/DataDimension/styles/Groups.style.js
