+++ packages/app/src/components/DimensionsPanel/Dialogs/Item.js
