+++ packages/app/src/components/DimensionsPanel/Dialogs/PeriodDimension/PeriodDimension.js
