+++ packages/app/src/components/DimensionsPanel/Dialogs/PeriodDimension/__tests__/PeriodDimension.spec.js
