+++ packages/app/src/components/DimensionsPanel/Dialogs/__tests__/DialogManager.spec.js
