+++ packages/app/src/components/DimensionsPanel/Dialogs/__tests__/SelectedItems.spec.js
