+++ packages/app/src/components/DimensionsPanel/Dialogs/styles/Dialog.css
