+++ packages/app/src/components/DimensionsPanel/Dialogs/styles/SelectedItems.style.js
