+++ packages/app/src/components/DimensionsPanel/List/DimensionList.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import DimensionItem from './DimensionItem';
 
-import { sGetDimensions } from '../../reducers/dimensions';
-import { sGetDimensionIdsFromLayout } from '../../reducers/ui';
+import { sGetDimensions } from '../../../reducers/dimensions';
+import { sGetDimensionIdsFromLayout } from '../../../reducers/ui';
 
 import { styles } from './styles/DimensionList.style';
