+++ packages/app/src/components/DimensionsPanel/List/OptionsButton.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import IconButton from '@material-ui/core/IconButton';
-import MoreHorizontalIcon from '../../assets/MoreHorizontalIcon';
+import MoreHorizontalIcon from '../../../assets/MoreHorizontalIcon';
 
 export const OptionsButton = ({ style, onClick }) => (
     <IconButton style={style} onClick={onClick}>
