+++ packages/app/src/components/DimensionsPanel/List/__tests__/RecommendedIcon.spec.js
