+++ packages/app/src/components/DimensionsPanel/List/styles/RecommendedIcon.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../modules/colors';
+import { colors } from '../../../../modules/colors';
 
 export const styles = {
     toolTip: {
