+++ packages/app/src/components/DimensionsPanel/styles/DimensionsPanel.style.js
